+++ src/shillelagh/backends/apsw/db.py
@@ -254,9 +254,10 @@ class Cursor:  # pylint: disable=too-many-instance-attributes
         """
         Build a ``DROP TABLE`` regexp.
         """
+        schema = re.escape(self.schema)
         regexp = re.compile(
-            rf"^\s*DROP\s+TABLE\s+(IF\s+EXISTS\s+)?"
-            rf'({self.schema}\.)?(?P<uri>(.*?)|(".*?"))\s*;?\s*$',
+            r"^\s*DROP\s+TABLE\s+(IF\s+EXISTS\s+)?"
+            rf'({schema}\.)?(?P<uri>(.*?)|(".*?"))\s*;?\s*$',
             re.IGNORECASE,
         )
         if match := regexp.match(operation):